# improve allocations in map serialization