--- a/io.go
+++ b/io.go
@@ -1,7 +1,10 @@
 package typegen
 
 import (
+	"fmt"
 	"io"
+
+	cid "github.com/ipfs/go-cid"
 )
 
 var (
@@ -42,6 +45,38 @@ func (cr *CborReader) ReadHeader() (byte, uint64, error) {
 
 func (cr *CborReader) SetReader(r io.Reader) {
 	cr.r = GetPeeker(r)
+}
+
+func (cr *CborReader) ReadCid(scratchBuf []byte) (cid.Cid, error) {
+	maj, extra, err := cr.ReadHeader()
+	if err != nil {
+		if err == io.EOF {
+			err = io.ErrUnexpectedEOF
+		}
+		return cid.Undef, err
+	}
+
+	if maj != MajTag {
+		return cid.Undef, fmt.Errorf("expected cbor type 'tag' in input")
+	}
+
+	if extra != 42 {
+		return cid.Undef, fmt.Errorf("expected tag 42")
+	}
+
+	if extra > 512 {
+		return cid.Undef, fmt.Errorf("header size too big for a cid")
+	}
+
+	_, err = io.ReadAtLeast(cr, scratchBuf[:extra], int(extra))
+	if err != nil {
+		if err == io.EOF {
+			err = io.ErrUnexpectedEOF
+		}
+		return cid.Undef, err
+	}
+
+	return bufToCid(scratchBuf[:extra])
 }
 
 var (